+++ StabilityMatrix.Avalonia/ViewModels/Dialogs/UpdateViewModel.cs
@@ -1,11 +1,15 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
+using System.Linq;
 using System.Net.Http;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using AsyncAwaitBestPractices;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using Semver;
+using StabilityMatrix.Avalonia.Languages;
 using StabilityMatrix.Avalonia.ViewModels.Base;
 using StabilityMatrix.Avalonia.Views.Dialogs;
 using StabilityMatrix.Core.Attributes;
@@ -23,6 +27,8 @@ public partial class UpdateViewModel : ContentDialogViewModelBase
     private readonly ISettingsManager settingsManager;
     private readonly IHttpClientFactory httpClientFactory;
     private readonly IUpdateHelper updateHelper;
+
+    private bool isLoaded;
 
     [ObservableProperty]
     private bool isUpdateAvailable;
@@ -42,6 +48,17 @@ public partial class UpdateViewModel : ContentDialogViewModelBase
     [ObservableProperty]
     private bool showProgressBar;
 
+    [ObservableProperty]
+    private string? currentVersionText;
+
+    [ObservableProperty]
+    private string? newVersionText;
+
+    [GeneratedRegex(
+        @"(##\s*(v[0-9]+\.[0-9]+\.[0-9]+)((?:\n|.)+?))(?=(##\s*v[0-9]+\.[0-9]+\.[0-9]+)|\z)"
+    )]
+    private static partial Regex RegexChangelog();
+
     public UpdateViewModel(
         ISettingsManager settingsManager,
         IHttpClientFactory httpClientFactory,
@@ -60,37 +77,91 @@ public partial class UpdateViewModel : ContentDialogViewModelBase
         updateHelper.StartCheckingForUpdates().SafeFireAndForget();
     }
 
-    public override async Task OnLoadedAsync()
+    /// <summary>
+    /// Formats changelog markdown including up to the current version
+    /// </summary>
+    internal static string? FormatChangelog(string markdown, SemVersion currentVersion)
+    {
+        var pattern = RegexChangelog();
+
+        var results = pattern
+            .Matches(markdown)
+            .Select(
+                m =>
+                    new
+                    {
+                        Block = m.Groups[1].Value.Trim(),
+                        Version = m.Groups[2].Value.Trim(),
+                        Content = m.Groups[3].Value.Trim()
+                    }
+            )
+            .ToList();
+
+        // Join all blocks until and excluding the current version
+        // If we're on a pre-release, include the current release
+
+        var currentVersionBlock = results.FindIndex(
+            x => x.Version == $"v{currentVersion.WithoutPrereleaseOrMetadata()}"
+        );
+
+        if (currentVersionBlock == -1)
+        {
+            return null;
+        }
+
+        var blocks = results
+            .Take(currentVersionBlock + (currentVersion.IsPrerelease ? 1 : 0))
+            .Select(x => x.Block)
+            .ToList();
+
+        return string.Join(Environment.NewLine + Environment.NewLine, blocks);
+    }
+
+    public async Task Preload()
     {
         if (UpdateInfo is null)
             return;
 
-        UpdateText =
-            $"Stability Matrix v{UpdateInfo.Version} is now available! You currently have v{Compat.AppVersion}. Would you like to update now?";
-
-        var client = httpClientFactory.CreateClient();
+        using var client = httpClientFactory.CreateClient();
         var response = await client.GetAsync(UpdateInfo.ChangelogUrl);
         if (response.IsSuccessStatusCode)
         {
-            ReleaseNotes = await response.Content.ReadAsStringAsync();
+            var changelog = await response.Content.ReadAsStringAsync();
 
             // Formatting for new changelog format
             // https://keepachangelog.com/en/1.1.0/
             if (UpdateInfo.ChangelogUrl.EndsWith(".md", StringComparison.OrdinalIgnoreCase))
             {
-                // Skip until the first occurrence of "##"
-                const string delimiter = "##";
-                var startIndex = ReleaseNotes.IndexOf(delimiter, StringComparison.Ordinal);
-                if (startIndex != -1)
-                {
-                    ReleaseNotes = ReleaseNotes[startIndex..];
-                }
+                ReleaseNotes =
+                    FormatChangelog(changelog, Compat.AppVersion)
+                    ?? "## Unable to format release notes";
             }
         }
         else
         {
             ReleaseNotes = "## Unable to load release notes";
         }
+    }
+
+    partial void OnUpdateInfoChanged(UpdateInfo? value)
+    {
+        CurrentVersionText = $"v{Compat.AppVersion}";
+        NewVersionText = $"v{value?.Version}";
+    }
+
+    public override async Task OnLoadedAsync()
+    {
+        if (!isLoaded)
+        {
+            await Preload();
+        }
+    }
+
+    /// <inheritdoc />
+    public override void OnUnloaded()
+    {
+        base.OnUnloaded();
+        isLoaded = false;
     }
 
     [RelayCommand]
@@ -102,7 +173,10 @@ public partial class UpdateViewModel : ContentDialogViewModelBase
         }
 
         ShowProgressBar = true;
-        UpdateText = $"Downloading update v{UpdateInfo.Version}...";
+        UpdateText = string.Format(
+            Resources.TextTemplate_UpdatingPackage,
+            Resources.Label_StabilityMatrix
+        );
         await updateHelper.DownloadUpdate(
             UpdateInfo,
             new Progress<ProgressReport>(report =>